pdf-worker/pyproject.toml
[project]
name = "pdf-worker"
version = "0.1.0"
description = "Searchable PDF builder worker for the archiver pipeline"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "httpx-sse>=0.4",
    "reportlab>=4.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
